Remove rdt.utils
